+++ ariadne_codegen/contrib/__init__.py
@@ -1,0 +1,4 @@
+from .extract_operations import ExtractOperationsPlugin
+from .shorter_results import ShorterResultsPlugin
+
+__all__ = ["ExtractOperationsPlugin", "ShorterResultsPlugin"]