+++ tests/main/clients/operations/expected_client/__init__.py
@@ -1,0 +1,62 @@
+from .async_base_client import AsyncBaseClient
+from .base_model import BaseModel, Upload
+from .c_mutation import CMutation
+from .c_query import CQuery
+from .c_subscription import CSubscription
+from .client import Client
+from .custom_operations import (
+    C_MUTATION_GQL,
+    C_QUERY_GQL,
+    C_SUBSCRIPTION_GQL,
+    GET_A_GQL,
+    GET_A_WITH_FRAGMENT_GQL,
+    GET_XYZ_GQL,
+)
+from .exceptions import (
+    GraphQLClientError,
+    GraphQLClientGraphQLError,
+    GraphQLClientGraphQLMultiError,
+    GraphQLClientHttpError,
+    GraphQlClientInvalidResponseError,
+)
+from .fragments import FragmentB, FragmentY
+from .get_a import GetA, GetAA, GetAAValueB
+from .get_a_with_fragment import (
+    GetAWithFragment,
+    GetAWithFragmentA,
+    GetAWithFragmentAValueB,
+)
+from .get_xyz import GetXYZ, GetXYZXyzTypeX, GetXYZXyzTypeY, GetXYZXyzTypeZ
+
+__all__ = [
+    "AsyncBaseClient",
+    "BaseModel",
+    "CMutation",
+    "CQuery",
+    "CSubscription",
+    "C_MUTATION_GQL",
+    "C_QUERY_GQL",
+    "C_SUBSCRIPTION_GQL",
+    "Client",
+    "FragmentB",
+    "FragmentY",
+    "GET_A_GQL",
+    "GET_A_WITH_FRAGMENT_GQL",
+    "GET_XYZ_GQL",
+    "GetA",
+    "GetAA",
+    "GetAAValueB",
+    "GetAWithFragment",
+    "GetAWithFragmentA",
+    "GetAWithFragmentAValueB",
+    "GetXYZ",
+    "GetXYZXyzTypeX",
+    "GetXYZXyzTypeY",
+    "GetXYZXyzTypeZ",
+    "GraphQLClientError",
+    "GraphQLClientGraphQLError",
+    "GraphQLClientGraphQLMultiError",
+    "GraphQLClientHttpError",
+    "GraphQlClientInvalidResponseError",
+    "Upload",
+]