+++ tests/main/clients/operations/expected_client/c_mutation.py
@@ -1,0 +1,10 @@
+from pydantic import Field
+
+from .base_model import BaseModel
+
+
+class CMutation(BaseModel):
+    const_mutation: int = Field(alias="constMutation")
+
+
+CMutation.model_rebuild()