+++ tests/main/clients/operations/expected_client/c_subscription.py
@@ -1,0 +1,10 @@
+from pydantic import Field
+
+from .base_model import BaseModel
+
+
+class CSubscription(BaseModel):
+    const_subscription: float = Field(alias="constSubscription")
+
+
+CSubscription.model_rebuild()