+++ tests/main/clients/operations/expected_client/fragments.py
@@ -1,0 +1,15 @@
+from pydantic import Field
+
+from .base_model import BaseModel
+
+
+class FragmentB(BaseModel):
+    value: str
+
+
+class FragmentY(BaseModel):
+    value_y: int = Field(alias="valueY")
+
+
+FragmentB.model_rebuild()
+FragmentY.model_rebuild()