+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Fixed `enums_module_name` option not being passed to generators.
 - Added additional base clients supporting the Open Telemetry tracing. Added `opentelemetry_client` config option.
 - Changed generated client's methods to pass `**kwargs` to base client's `execute` and `execute_ws` methods (breaking change for custom base clients).
+- Added `operation_definition` argument to `generate_client_method` plugin hook.
+- Added `ExtractOperationsPlugin` that extracts operation strings from client methods to separate module.
 
 
 ## 0.9.0 (2023-09-11)